+++ src/main/java/Dao/Manipulation.java
@@ -6,7 +6,6 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
-import java.time.LocalDateTime;
 
 /**
  * Entity manipulation
@@ -22,11 +21,11 @@ public class Manipulation {
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     //name of manipulation
     @Column(name = "manipulation_title", nullable = false)
-    private String  title;
+    private String title;
 
     //type of manipulation
     @Column(name = "manipulation_type", nullable = false)
